Bit of a large re-structure
Hesitate to say 'refactor' here because I tried not to actually change
any of the functionality yet. This was mostly just to break things out
of the Saltwalk class which didn't really need to be a class at
all, as there was only a single bit of state (self.pending_jobs) that
was actually being modified by any of the private methods in the
class.
I also added a simple contextmanager to handle rhnSQL transactions that
were potentially being rolled back in two different ways (one inside the
method, the other outside the method in 'run').
Creating a named logger now as well.